requirements.txt
about-time==3.1.1
alive-progress==2.1.0
appdirs==1.4.4
APScheduler==3.6.3
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
fake-headers==1.0.2
grapheme==0.6.0
greenlet==1.1.2
html5lib==1.1
idna==3.3
progress==1.6
progressbar2==4.0.0
pydantic==1.9.0
PySocks==1.7.1
python-dotenv==0.20.0
python-telegram-bot==13.11
python-utils==3.1.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-tor==1.3
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
telegram-send==0.25
tornado==6.1
tqdm==4.62.3
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
webencodings==0.5.1
